+++ Website/next.config.ts
@@ -1,10 +1,10 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  outputFileTracingRoot: __dirname,
   typescript: {
     ignoreBuildErrors: false,
   },
+  outputFileTracingRoot: __dirname,
 };
 
 export default nextConfig;